+++ packages/lumx-react/src/components/mosaic/Mosaic.tsx
@@ -1,14 +1,16 @@
 import React from 'react';
 
-import { AspectRatio, FocusPoint, Theme, Thumbnail } from '@lumx/react';
+import { AspectRatio, Theme, Thumbnail, ThumbnailProps } from '@lumx/react';
 import { COMPONENT_PREFIX } from '@lumx/react/constants';
 import { GenericProps, getRootClassName, handleBasicClasses } from '@lumx/react/utils';
 import classNames from 'classnames';
 import take from 'lodash/take';
 
-interface MosaicElement {
+interface MosaicElement extends Omit<ThumbnailProps, 'aspectRatio' | 'fillHeight' | 'theme' | 'image'> {
+    /** @deprecated Use `image` instead. */
     url: string;
-    focusPoint?: FocusPoint;
+    /** @todo breaking change removing the `url` prop and making the `image` required. */
+    image?: ThumbnailProps['image'];
     onClick?(index: number): void;
 }
 
@@ -38,33 +40,34 @@ const CLASSNAME: string = getRootClassName(COMPONENT_NAME);
 const DEFAULT_PROPS: DefaultPropsType = {
     theme: Theme.light,
 };
-const Mosaic: React.FC<MosaicProps> = ({ className, theme = DEFAULT_PROPS.theme, thumbnails, ...props }) => (
+const Mosaic: React.FC<MosaicProps> = ({ className, theme = DEFAULT_PROPS.theme, thumbnails, ...forwardedProps }) => (
     <div
+        {...forwardedProps}
         className={classNames(className, handleBasicClasses({ prefix: CLASSNAME, theme }), {
+            [`${CLASSNAME}--has-1-thumbnail`]: thumbnails?.length === 1,
             [`${CLASSNAME}--has-2-thumbnails`]: thumbnails?.length === 2,
             [`${CLASSNAME}--has-3-thumbnails`]: thumbnails?.length === 3,
             [`${CLASSNAME}--has-4-thumbnails`]: thumbnails?.length >= 4,
         })}
-        {...props}
     >
         <div className={`${CLASSNAME}__wrapper`}>
-            {take(thumbnails, 4).map((thumbnail, index) => {
+            {take(thumbnails, 4).map(({ url, image, onClick, ...thumbnail }, index) => {
                 const handleClick = () => {
-                    if (thumbnail.onClick) {
-                        thumbnail.onClick(index);
+                    if (onClick) {
+                        onClick(index);
                     }
                 };
 
                 return (
                     <div key={index} className={`${CLASSNAME}__thumbnail`}>
                         <Thumbnail
-                            tabIndex="0"
-                            onClick={handleClick}
+                            {...thumbnail}
+                            theme={theme}
+                            image={image ?? url}
                             aspectRatio={AspectRatio.free}
                             fillHeight
-                            image={thumbnail.url}
-                            theme={theme}
-                            focusPoint={thumbnail.focusPoint}
+                            tabIndex={onClick && '0'}
+                            onClick={handleClick}
                         />
 
                         {thumbnails.length > 4 && index === 3 && (